+++ electron/main/utils/common.ts
@@ -1,0 +1,11 @@
+export function randomInt(min: number, max: number): number {
+  return Math.floor(Math.random() * (max - min + 1)) + min
+}
+
+export async function sleep(ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms))
+}
+
+export function isDev() {
+  return process.env.NODE_ENV === 'development'
+}
