+++ electron/main/utils/types.ts
@@ -1,0 +1,6 @@
+export type PromiseReturnType<T> = T extends Promise<infer U> ? U : T
+
+export type RequiredWith<BaseType, Keys extends keyof BaseType> = Required<
+  Pick<BaseType, Keys>
+> &
+  Omit<BaseType, Keys>
